Information from blocks sent from client is now properly handled. Client also receives updated info but this still needs implementation into ui

+++ digital_twin_server/src/main/java/com/nhlstenden/groep3/digital_twin/DTO/BlockDTO.java
@@ -30,7 +30,7 @@ public class BlockDTO {
 
     public BlockDTO(){}
 
-    public BlockDTO(Long id, BlockType blockType, Long mapId, Polygon geometry, float height, float area, float volume, float calculated_cost, float calculated_yield, float calculated_residents) {
+    public BlockDTO(Long id, BlockType blockType, Long mapId, Polygon geometry, float height, float area, float volume, float calculated_cost, float calculated_yield, int calculated_residents) {
         this.id = id;
         this.blockType = blockType;
         this.mapId = mapId;
